+++ R/afs_capewide_pup.R
@@ -15,17 +15,23 @@
 #'
 #' afs_cwp_totals_bylocation: group by season_name and location;
 #' return count mean, variance, SD, and min date.
-#' Var and SD follow CS/AMLR-specific rules about when it must be NA
+#' Var and SD follow CS/AMLR-specific rules about when it must be NA.
+#' census_date_min is the earliest census date for that season/location group.
 #'
 #' afs_cwp_totals: after running afs_cwp_totals_bylocation,
 #' group by season name and return counts and standard deviations
+#'
+#' Note that none of these functions do any rounding. To round values,
+#' for instance to get a whole number for pup counts,
+#' use either \code{\link[base]{round}}
+#' or \code{\link[amlrDatabases]{round_logical}}
 #'
 #' @return
 #' An ungrouped data frame with the summarized AFS Capewide Pup Census data.
 #' The summary types are described in 'Details'.
 #'
 #' @examplesIf FALSE
-#' # Only will run if on SWFSC network
+#' # Not run; examples only will run if on SWFSC network
 #' x <- tbl_vCensus_AFS_Capewide_Pup(amlr_dbConnect("***REMOVED***_Test"))
 #'
 #' afs_cwp_single(x[x$season_name == "2016/17", ])
@@ -52,30 +58,16 @@ afs_cwp_single <- function(x) { #}, by.observer) {
     group_by(season_name, census_afs_capewide_pup_sort, location) %>%
     arrange(observer) %>% #so that collapsed data are always in the same order
     summarise(n_records = n(),
-              count_mean = round(mean(pup_count), 1),
+              count_mean = mean(pup_count),
               count_range = diff(range(pup_count)),
               count_range_perc_diff = if_else(
-                count_mean == 0, 0,
-                round(count_range / count_mean * 100, 2)),
+                count_mean == 0, 0, count_range / count_mean * 100),
               observers = paste(observer, collapse = "; "),
               counts = paste(pup_count, collapse = "; "),
               exclude_count = paste(as.integer(exclude_count), collapse = "; "),
               notes = paste(census_notes, collapse = "; "),
               counts_live = paste(pup_live_count, collapse = "; "),
               counts_dead = paste(pup_dead_count, collapse = "; "),
-              # counts = paste(paste(observer, pup_count, sep = ": "),
-              #                collapse = "; "),
-              # exclude_count = paste(as.integer(exclude_count), collapse = "; "),
-              # notes = paste(na.omit(
-              #   if_else(
-              #     is.na(census_notes), NA_character_,
-              #     paste(observer, census_notes, sep = ": ")
-              #   )
-              # ), collapse = "; "),
-              # counts_live = paste(paste(observer, pup_dead_count, sep = ": "),
-              #                     collapse = "; "),
-              # counts_dead = paste(paste(observer, pup_live_count, sep = ": "),
-              #                     collapse = "; "),
               .groups = "drop") %>%
     arrange(census_afs_capewide_pup_sort) %>%
     select(-c(census_afs_capewide_pup_sort))
@@ -99,14 +91,14 @@ afs_cwp_totals_bylocation <- function(x) {
     summarise(num_records = n(),
               count_loc_mean = mean(pup_count),
               count_loc_var = var(pup_count),
-              date_min = min(census_date),
+              census_date_min = min(census_date),
               research_program = unique(research_program),
               .groups = "drop") %>%
-    mutate(study_beach_count = between(date_min,
+    mutate(study_beach_count = between(census_date_min,
                                        ymd("2008-07-01"), ymd("2011-07-01")) &
              location %in% c("Copihue", "Maderas", "Cachorros", "Chungungo"),
            count_loc_var = case_when(
-             date_min < as.Date("2008-07-01") ~ NA_real_,
+             census_date_min < as.Date("2008-07-01") ~ NA_real_,
              study_beach_count ~ NA_real_,
              .default = count_loc_var
            ),
@@ -122,7 +114,7 @@ afs_cwp_totals <- function(x, x.bylocation = FALSE) {
     stopifnot(identical(
       c("season_name", "location",
         "num_records", "count_loc_mean", "count_loc_sd", "count_loc_var",
-        "date_min", "research_program"),
+        "census_date_min", "research_program"),
       names(x)
     ))
     x
@@ -132,10 +124,10 @@ afs_cwp_totals <- function(x, x.bylocation = FALSE) {
 
   y %>%
     group_by(season_name) %>%
-    summarise(count_mean = round(sum(count_loc_mean), 0),
-              count_var = if_else(min(date_min) < as.Date("2011-07-01"),
+    summarise(count_mean = sum(count_loc_mean),
+              count_var = if_else(min(census_date_min) < as.Date("2011-07-01"),
                                   NA_real_, sum(count_loc_var, na.rm = TRUE)),
-              count_sd = round(sqrt(count_var), 0),
+              count_sd = sqrt(count_var),
               research_program = unique(research_program),
               .groups = "drop") %>%
     select(-count_var)